pyproject.toml
[tool.poetry]
name = "rsa_generator"
version = "0.1.0"
description = ""
authors = ["Mio Lau <liurusi.101@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
Nuitka = "^0.7.7"
click = "^8.1.2"
pycryptodome = "^3.14.1"

[tool.poetry.dev-dependencies]
ipython = "^8.2.0"
pyinstaller = "^4.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true